fix react-env template
