+++ scopes/react/react/react.templates.ts
@@ -1,6 +1,6 @@
 import { ComponentTemplate } from '@teambit/generator';
 import { reactButton } from './templates/react-button';
 import { reactComponent } from './templates/react-component';
-import { reactExtension } from './templates/react-extension';
+import { customReactEnv } from './templates/custom-react-env';
 
-export const componentTemplates: ComponentTemplate[] = [reactButton, reactComponent, reactExtension];
+export const componentTemplates: ComponentTemplate[] = [reactButton, reactComponent, customReactEnv];
