+++ scopes/react/react/templates/custom-react-env/files/doc.ts
@@ -1,0 +1,11 @@
+export function docFile() {
+  return `---
+labels: ['extension', 'react', 'env', 'environment']
+description: 'A customized extension for React environment.'
+---
+
+This is a customized extension for React, based of the [React Aspect](https://bit.dev/teambit/react/react).
+
+Explain here the modified configurations and tools applied.
+`;
+}
