+++ scopes/react/react/templates/custom-react-env/files/index.ts
@@ -1,0 +1,8 @@
+import { GeneratorContext } from '@teambit/generator';
+
+export function indexFile({ componentNameCamelCase: Name, componentName }: GeneratorContext) {
+  return `import { ${Name}Extension } from './${componentName}.extension';
+export { ${Name}Extension };
+export default ${Name}Extension;
+`;
+}
